Add files via upload
Flow to load battery based on the Tibber prices. 
The current version loads in the 4 cheapest hours in 48 hours the battery to 50%. 
When the price is higher than 4th lowest price +20% the battery is used. 
Between 4th lowest price und 4th lowest +19% the battery stays on the current soc.

Todo: Dynamic target soc based on temperature and expected solar yield.